Add support for Darius's passive.
This required a surprising amount of work, including
adding support for passives in the first place (or other
champion effects).  I also had to create a system of buffs
and DOTs as well as did some refactoring of both cooldowns
(finding a bug in how I was applying the AA cooldown buff
to the *target* instead of the source) as well as the damage
log to support recording different types of damage.

Also moved buffs out of World and onto Mobs for now.