Fix round_robin.dart to not treat < 1hp as dead.
At least in the current simulation it's possible to be
alive and have less than 1 hp, however round_robin was not
respecting that and thus getting incorrect results.

This fixes that bug as well as adds a simple_duel script
which I used heavily in comparing these many small differences.