More work to support modeling Jungle clears.
This required adding support for Hunter's Talisman
Which required implementing mana regeneration
as well as mpRegen stat modifiers.
Which caused me to flesh out hpRegen stat modifiers.

This also required more accessor methods on SpellEffects
which prompted separation of spell methods from effects
methods, as well as filling in cooldowns on various
spells.

Much of this code is not yet well tested, but I'm commiting
this as is.  I also have my clear simulator script yet to
land on a future commit.